Fix bug in error annotation incorrectly mapping buffer marker to jump list line
The buffer markers only contains the markers for that buffer. Jump list/state from a jump buffer contains the entire jump list/state, which may contain jump locations to several different buffers. In other words, the i-th line from the jump list does not necessarily correspond to the i-th buffer marker.

In order to get the correct jump line, we need to go through the jump list, skip over the ones not correspnding to the buffer currently being rendered, and then use its index_into_marker_array to get the corresponding marker. There may be better ways of doing this